Permissions back-end

+++ app/Http/Controllers/PostController.php
@@ -69,6 +69,7 @@ class PostController extends Controller
      */
     public function store(StorePostRequest $request)
     {
+        $this->authorize('posts.create');
         if ($request->hasFile('thumbnail')) {
             $filename = $request->file('thumbnail')->getClientOriginalName();
             info($filename);
@@ -86,6 +87,7 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        $this->authorize('posts.update');
         return new PostResource($post);
     }
 
@@ -109,6 +111,7 @@ class PostController extends Controller
      */
     public function update(StorePostRequest $request, Post $post)
     {
+        $this->authorize('posts.update');
         $post->update($request->validated());
 
         return new PostResource($post);
@@ -122,6 +125,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('posts.delete');
         $post->delete();
 
         return response()->noContent();
